+++ test/sandbox/test/generators/slim_generator_test.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "rails/generators/slim/component_generator"
+require "generators/view_component/slim/slim_generator"
 
 Rails.application.load_generators
 
 class SlimGeneratorTest < Rails::Generators::TestCase
-  tests Slim::Generators::ComponentGenerator
+  tests ViewComponent::Generators::SlimGenerator
   destination Dir.mktmpdir
   setup :prepare_destination